Reduce number of Windows warning C4365

+++ test/libsrcml/testsuite/test_srcml_archive_write_open.cpp
@@ -38,7 +38,7 @@
 
 extern "C" {
     int write_callback(void * context, const char * buffer, int len) {
-        return (int)fwrite(buffer, 1, len, (FILE*)context);
+        return (int)fwrite(buffer, 1, (size_t)len, (FILE*)context);
 
     }
